Refactor knowledge and session repositories
